Corrected the Download Excel button wn and updated the hierarchy

+++ rails_root/app/controllers/survey_profiles_controller.rb
@@ -127,20 +127,20 @@ class SurveyProfilesController < ApplicationController
 
   def claim_invitation
     temporary_invitation_session_var = session[:invitation]
-
-    if temporary_invitation_session_var && temporary_invitation_session_var['expiration'] > Time.now
+  
+    if temporary_invitation_session_var
       invitation = Invitation.find_by(id: temporary_invitation_session_var['from'])
+      
       if invitation
         sharecode_from_invitation = invitation.parent_response.share_code
-        # survey_profile = SurveyProfile.find_by(user_id: session[:userinfo]['sub'])
         new_response_to_fill = SurveyResponse.create(profile: @survey_profile, share_code: sharecode_from_invitation)
         invitation.update(claimed_by_id: @survey_profile.id, response_id: new_response_to_fill.id)
         redirect_to edit_survey_response_path(new_response_to_fill)
         return true
       end
     end
-
-    session.delete(:invitation)
+  
+    session.delete(:invitation) if temporary_invitation_session_var.nil?
     false
-  end
+  end  
 end
